refactor: move nutrition helper functions to utils
